refactor: Rewrite the code to use pyflowlauncher
